requirements.txt
opencv-python==4.10.0.84
pillow==10.4.0
pandas==2.2.2
numpy==2.1.1
